Fix initializing indices with more than one document
The testing was done with an index with a single document. It was incorrectly
thought that the `created` field counts the number of indices created, but it's
actually the number of documents, which is likely to be larger than 1.

The `created` was changed to `total`, because when something fails and the same
operation is run again, the reindex will _update_, not _create_ the documents.

INF-1553