Throttle reindexing
Information about the settings can be found in [ES docs][1].

`wait_for_active_shards` changes from the default value of 1 to `all`. This
means that readiness of replications shards will also be waited for, before
continuing.

`requests_per_second` is changed from the default -1, to 500 which enables
throttling at the specified rate, further explained in the doc.

`timeout` is changed from the default 1 minute, because these reindexings can
take a while.

INF-1553

[1]: https://www.elastic.co/guide/en/elasticsearch/reference/current/docs-reindex.html#_url_parameters_3