Use size instead of count for rollover limit
I originally thought that shard size was unreliable metric to use for the
rollover because of the section in [Managing Elasticsearch time-based indices
efficiently][1]:

> [..] shard size is a less reliable measure because ongoing merges can produce
significant temporary growth in shard size which disappears as soon as a merge
is completed.  Five primary shards, each in the process of merging to one 5GB
shard, would temporarily increase the index size by 25GB!  The doc count, on
the other hand grows predictably.

But according to [this comment][2], the `max_size` option has been
[implemented][3] using more reliable metrics and can be used here. This is
better for us, because what we were trying to aim for with the `max_docs`
option was a specific index/shard size.

INF-1553

[1]: https://www.elastic.co/blog/managing-time-based-indices-efficiently
[2]: https://github.com/elastic/elasticsearch/issues/27004#issuecomment-339408669
[3]: https://github.com/elastic/elasticsearch/pull/27160