Accommodate for snapshot restores
When a log cluster is created from scratch and its logs are restored from
a snapshot (without restoring aliases), then we get to a state where fluentd
writes to a suffixless (uninitialized) index, while the cluster also has
indices with suffixes. For initialization to work nicely in that scenario, we
have to reindex the data from the suffixless (uninitialized) index into the
index with the largest suffix and then point the new alias to that index.

INF-1553