Add tests rollover index initialization
The `discovery.type` configuration value is specified in the README and used in
the Jenkinsfile to start the Elasticsearch node in _development_ node and
disable some bootstrap checks. In particular, the [Maximum map count check][1]
is disabled, which would fail in our Kubernetes CI environment. The
configuration value is defined with `args` and not with env vars, because
[prior to version 1.8][2], Kubernetes doesn't support environment variables
with periods ('.') in the name.

INF-1553

[1]: https://www.elastic.co/guide/en/elasticsearch/reference/6.2/_maximum_map_count_check.html
[2]: https://github.com/kubernetes/kubernetes/pull/48986