Continue with rollover initialization on reindexing failure
For some of our indices that get a lot of messages, the reindexing doesn't
complete within an hour. This leaves the index blocked for writes and doesn't
create the alias.

As the test shows, this change sacrifices data within the original index to
ensure that the initialization succeeds and that the index allows new data to
flow in.

INF-1906